backend/requirements.txt
# ============================================================
# Intelligent Real-Time Speed Limit Sign Detection System
# Requirements File
# ============================================================

# Core ML & Vision
ultralytics>=8.0.0          # YOLOv8 object detection framework
opencv-python>=4.8.0        # Video/image processing
opencv-contrib-python>=4.8.0 # Extra OpenCV modules

# OCR
pytesseract>=0.3.10         # Python wrapper for Tesseract OCR
Pillow>=9.0.0               # Image processing for OCR pipeline

# Deep Learning Backend
torch>=2.0.0                # PyTorch (CPU version by default)
torchvision>=0.15.0

# Tracking (DeepSORT)
deep-sort-realtime>=1.3.2   # DeepSORT multi-object tracker

# Data & Logging
numpy>=1.23.0
pandas>=1.5.0               # CSV logging and analysis

# Visualization
matplotlib>=3.6.0           # Plotting detection statistics

# Utilities
pyyaml>=6.0                 # Config file parsing
tqdm>=4.64.0                # Progress bars
colorlog>=6.7.0             # Colored terminal logging
python-dotenv>=1.0.0        # Environment variable management

# Optional: GPU acceleration (uncomment if CUDA available)
# torch>=2.0.0+cu118
# torchvision>=0.15.0+cu118

FastAPI>=0.95.0             # For API development (optional)
uvicorn>=0.34.0            # ASGI server for FastAPI (optional)

